+++ demo/stories/ghost/Ghost.tsx
@@ -1,16 +1,11 @@
 import cloneDeep from 'lodash/cloneDeep';
 
-import {MarkdownEditorView, logger, markupToolbarConfigs, useMarkdownEditor} from '../../../src';
+import {MarkdownEditorView, markupToolbarConfigs, useMarkdownEditor} from '../../../src';
 import {PlaygroundLayout} from '../../components/PlaygroundLayout';
+import {useLogs} from '../../hooks/useLogs';
 
 import {initialMdContent} from './content';
 import {ghostPopupExtension, ghostPopupToolbarItem} from './ghostExtension';
-
-logger.setLogger({
-    metrics: console.info,
-    action: (data) => console.info(`Action: ${data.action}`, data),
-    ...console,
-});
 
 const mToolbarConfig = cloneDeep(markupToolbarConfigs.mToolbarConfig);
 mToolbarConfig.unshift([ghostPopupToolbarItem]);
@@ -20,6 +15,8 @@ export const Ghost = () => {
         initial: {markup: initialMdContent, mode: 'markup'},
         markupConfig: {extensions: [ghostPopupExtension]},
     });
+
+    useLogs(editor.logger);
 
     return (
         <PlaygroundLayout