+++ src/core/Editor.ts
@@ -3,6 +3,7 @@ import {EditorState} from 'prosemirror-state';
 import {EditorView} from 'prosemirror-view';
 
 import type {CommonEditor, ContentHandler, MarkupString} from '../common';
+import {Logger2} from '../logger';
 
 import type {ActionsManager} from './ActionsManager';
 import {WysiwygContentHandler} from './ContentHandler';
@@ -18,6 +19,7 @@ import type {Parser} from './types/parser';
 import type {Serializer} from './types/serializer';
 import {bindActions} from './utils/actions';
 import {convertDynamicModifiersConfigs} from './utils/dynamicModifiers';
+import {LoggerFacet} from './utils/logger';
 import {logTransactionMetrics} from './utils/metrics';
 
 type OnChange = (editor: WysiwygEditor) => void;
@@ -45,6 +47,7 @@ export type WysiwygEditorOptions = {
     onDocChange?: OnChange;
     /** @internal Modifiers adjust the parser and serializer */
     modifiers?: DynamicModifiers[];
+    logger?: Logger2.ILogger;
 };
 
 export class WysiwygEditor implements CommonEditor, ActionStorage {
@@ -89,6 +92,7 @@ export class WysiwygEditor implements CommonEditor, ActionStorage {
         onChange,
         onDocChange,
         modifiers,
+        logger = new Logger2(),
     }: WysiwygEditorOptions) {
         const dynamicModifiersConfig = modifiers
             ? convertDynamicModifiersConfigs(modifiers)
@@ -112,13 +116,19 @@ export class WysiwygEditor implements CommonEditor, ActionStorage {
             plugins,
             rawActions,
             actions,
-        } = ExtensionsManager.process(extensions, {
-            // "breaks" option only affects the renderer, but not the parser
-            mdOpts: {html: allowHTML, linkify, breaks: true, preset: mdPreset},
-            linkifyTlds,
-            pmTransformers,
-            dynamicModifiers,
-        });
+        } = ExtensionsManager.process(
+            extensions,
+            {
+                // "breaks" option only affects the renderer, but not the parser
+                mdOpts: {html: allowHTML, linkify, breaks: true, preset: mdPreset},
+                linkifyTlds,
+                pmTransformers,
+                dynamicModifiers,
+            },
+            logger,
+        );
+
+        plugins.unshift(LoggerFacet.of(logger));
 
         const state = EditorState.create({
             schema,
@@ -141,7 +151,7 @@ export class WysiwygEditor implements CommonEditor, ActionStorage {
                 if (tr.docChanged) {
                     thisOnDocChange();
                 }
-                logTransactionMetrics(tr);
+                logTransactionMetrics(tr, logger);
             },
         });
         this.#actions = actions.setActions(