+++ src/extensions/behavior/Autocomplete/handler.ts
@@ -1,5 +1,5 @@
 import {capitalize, isString} from '../../../lodash';
-import {logger} from '../../../logger';
+import {type Logger2, globalLogger} from '../../../logger';
 
 import type {AutocompleteAction, AutocompleteHandler, AutocompleteItem} from './types';
 
@@ -16,11 +16,12 @@ type MainHandlerLogOptions =
       };
 
 export class MainHandler implements Required<AutocompleteHandler> {
-    private static log(opts: MainHandlerLogOptions) {
+    private static log(logger: Logger2.ILogger, opts: MainHandlerLogOptions) {
         const prefix = '[Autocomplete]';
 
         if (opts.event === 'destroy') {
             logger.log(`${prefix} Destroy`);
+            globalLogger.log(`${prefix} Destroy`);
             return;
         }
 
@@ -35,46 +36,51 @@ export class MainHandler implements Required<AutocompleteHandler> {
         ].join(' ');
 
         logger.log(msg);
+        globalLogger.log(msg);
     }
 
     private readonly config: MainHandlerConfig;
+    private readonly logger: Logger2.ILogger;
 
-    constructor(config: MainHandlerConfig) {
+    constructor(config: MainHandlerConfig, logger: Logger2.ILogger) {
         this.config = config;
+        this.logger = logger.nested({
+            module: 'autocomplete-main-handler',
+        });
     }
 
     onOpen(action: AutocompleteAction): boolean {
         const item = this.getItem(action);
-        MainHandler.log({event: 'open', action, item});
+        MainHandler.log(this.logger, {event: 'open', action, item});
         return item?.handler.onOpen?.(action) ?? false;
     }
 
     onClose(action: AutocompleteAction): boolean {
         const item = this.getItem(action);
-        MainHandler.log({event: 'close', action, item});
+        MainHandler.log(this.logger, {event: 'close', action, item});
         return item?.handler.onClose?.(action) ?? false;
     }
 
     onArrow(action: AutocompleteAction): boolean {
         const item = this.getItem(action);
-        MainHandler.log({event: 'arrow', action, item});
+        MainHandler.log(this.logger, {event: 'arrow', action, item});
         return item?.handler.onArrow?.(action) ?? false;
     }
 
     onEnter(action: AutocompleteAction): boolean {
         const item = this.getItem(action);
-        MainHandler.log({event: 'enter', action, item});
+        MainHandler.log(this.logger, {event: 'enter', action, item});
         return item?.handler.onEnter?.(action) ?? false;
     }
 
     onFilter(action: AutocompleteAction): boolean {
         const item = this.getItem(action);
-        MainHandler.log({event: 'filter', action, item});
+        MainHandler.log(this.logger, {event: 'filter', action, item});
         return item?.handler.onFilter?.(action) ?? false;
     }
 
     onDestroy(): void {
-        MainHandler.log({event: 'destroy'});
+        MainHandler.log(this.logger, {event: 'destroy'});
         for (const item of this.config) {
             item.handler.onDestroy?.();
         }